+++ antibytes-dependency/src/main/kotlin/tech/antibytes/gradle/dependency/Version.kt
@@ -9,6 +9,7 @@ package tech.antibytes.gradle.dependency
 import tech.antibytes.gradle.dependency.version.Android
 import tech.antibytes.gradle.dependency.version.Google
 import tech.antibytes.gradle.dependency.version.Gradle
+import tech.antibytes.gradle.dependency.version.Js
 import tech.antibytes.gradle.dependency.version.Jvm
 import tech.antibytes.gradle.dependency.version.Kotlin
 import tech.antibytes.gradle.dependency.version.Npm
@@ -29,15 +30,7 @@ object Version {
 
     val google = Google
 
-    val js = JS
+    val js = Js
 
     val npm = Npm
-
-    object JS {
-        /**
-         * [KotlinNodeJS](https://github.com/Kotlin/kotlinx-nodejs)
-         */
-        const val nodeJs = "0.0.7"
-        const val node = "18.9.0"
-    }
 }
