+++ antibytes-dependency/src/main/kotlin/tech/antibytes/gradle/dependency/packages/kmp/AtomicFu.kt
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
+ *
+ * Use of this source code is governed by Apache License, Version 2.0
+ */
+
+package tech.antibytes.gradle.dependency.packages.kmp
+
+import tech.antibytes.gradle.dependency.Version
+
+object AtomicFu {
+    const val common = "org.jetbrains.kotlinx:atomicfu:${Version.kotlin.atomicFu}"
+}
