+++ antibytes-dependency/src/main/kotlin/tech/antibytes/gradle/dependency/packages/kmp/Coroutines.kt
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
+ *
+ * Use of this source code is governed by Apache License, Version 2.0
+ */
+
+package tech.antibytes.gradle.dependency.packages.kmp
+
+import tech.antibytes.gradle.dependency.Version
+
+object Coroutines {
+    const val common = "org.jetbrains.kotlinx:kotlinx-coroutines-core:${Version.kotlin.coroutines}"
+    const val android = "org.jetbrains.kotlinx:kotlinx-coroutines-android:${Version.kotlin.coroutines}"
+    const val js = "org.jetbrains.kotlinx:kotlinx-coroutines-core-js:${Version.kotlin.coroutines}"
+}
