+++ antibytes-dependency/src/main/kotlin/tech/antibytes/gradle/dependency/packages/kmp/Koin.kt
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
+ *
+ * Use of this source code is governed by Apache License, Version 2.0
+ */
+
+package tech.antibytes.gradle.dependency.packages.kmp
+
+import tech.antibytes.gradle.dependency.Version
+
+object Koin {
+    const val core = "io.insert-koin:koin-core:${Version.kotlin.koin}"
+    const val android = "io.insert-koin:koin-android:${Version.kotlin.koin}"
+    const val androidAppCompact = "io.insert-koin:koin-android-compat:${Version.kotlin.koin}"
+    const val jetpackWorkmanager = "io.insert-koin:koin-androidx-workmanager:${Version.kotlin.koin}"
+    const val jetpackNavigation = "io.insert-koin:koin-androidx-navigation:${Version.kotlin.koin}"
+    const val jetpackCompose = "io.insert-koin:koin-androidx-compose:${Version.kotlin.koin}"
+}
