+++ antibytes-dependency/src/main/kotlin/tech/antibytes/gradle/dependency/packages/kmp/Ktor.kt
@@ -1,0 +1,88 @@
+/*
+ * Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
+ *
+ * Use of this source code is governed by Apache License, Version 2.0
+ */
+
+package tech.antibytes.gradle.dependency.packages.kmp
+
+import tech.antibytes.gradle.dependency.Version
+
+object Ktor {
+    val common = Common
+
+    object Common {
+        const val client = "io.ktor:ktor-client-core:${Version.kotlin.ktor}"
+        const val json = "io.ktor:ktor-serialization-kotlinx-json:${Version.kotlin.ktor}"
+        const val xml = "io.ktor:ktor-serialization-kotlinx-xml:${Version.kotlin.ktor}"
+        const val cbor = "io.ktor:ktor-serialization-kotlinx-cbor:${Version.kotlin.ktor}"
+        const val websockets = "io.ktor:ktor-client-websockets:${Version.kotlin.ktor}"
+        const val contentNegotiation = "io.ktor:ktor-client-content-negotiation:${Version.kotlin.ktor}"
+    }
+
+    val jvm = Jvm
+
+    object Jvm {
+        const val client = "io.ktor:ktor-client-java:${Version.kotlin.ktor}"
+        const val jetty = "io.ktor:ktor-client-jetty:${Version.kotlin.ktor}"
+        const val apache = "io.ktor:ktor-client-apache:${Version.kotlin.ktor}"
+        const val cio = "io.ktor:ktor-client-cio:${Version.kotlin.ktor}"
+        const val json = "io.ktor:ktor-serialization-kotlinx-json-jvm:${Version.kotlin.ktor}"
+        const val xml = "io.ktor:ktor-serialization-kotlinx-xml-jvm:${Version.kotlin.ktor}"
+        const val cbor = "io.ktor:ktor-serialization-kotlinx-cbor-jvm:${Version.kotlin.ktor}"
+        const val websockets = "io.ktor:ktor-client-websockets-jvm:${Version.kotlin.ktor}"
+        const val contentNegotiation = "io.ktor:ktor-client-content-negotiation-jvm:${Version.kotlin.ktor}"
+    }
+
+    val android = Android
+
+    object Android {
+        const val client = "io.ktor:ktor-client-android:${Version.kotlin.ktor}"
+        const val okhttp = "io.ktor:ktor-client-okhttp:${Version.kotlin.ktor}"
+        const val json = "io.ktor:ktor-serialization-kotlinx-json-android:${Version.kotlin.ktor}"
+        const val xml = "io.ktor:ktor-serialization-kotlinx-xml-android:${Version.kotlin.ktor}"
+        const val cbor = "io.ktor:ktor-serialization-kotlinx-cbor-android:${Version.kotlin.ktor}"
+        const val websockets = "io.ktor:ktor-client-websockets-android:${Version.kotlin.ktor}"
+        const val contentNegotiation = "io.ktor:ktor-client-content-negotiation-android:${Version.kotlin.ktor}"
+    }
+
+    val ios = Ios
+
+    object Ios {
+        const val client = "io.ktor:ktor-client-darwin:${Version.kotlin.ktor}"
+        const val json = "io.ktor:ktor-serialization-kotlinx-json-native:${Version.kotlin.ktor}"
+        const val xml = "io.ktor:ktor-serialization-kotlinx-xml-native:${Version.kotlin.ktor}"
+        const val cbor = "io.ktor:ktor-serialization-kotlinx-cbor-native:${Version.kotlin.ktor}"
+        const val websockets = "io.ktor:ktor-client-websockets-native:${Version.kotlin.ktor}"
+        const val contentNegotiation = "io.ktor:ktor-client-content-negotiation-native:${Version.kotlin.ktor}"
+    }
+
+    val js = JS
+
+    object JS {
+        const val client = "io.ktor:ktor-client-js:${Version.kotlin.ktor}"
+        const val serialization = "io.ktor:ktor-client-serialization-js:${Version.kotlin.ktor}"
+        const val json = "io.ktor:ktor-serialization-kotlinx-json-js:${Version.kotlin.ktor}"
+        const val xml = "io.ktor:ktor-serialization-kotlinx-xml-js:${Version.kotlin.ktor}"
+        const val cbor = "io.ktor:ktor-serialization-kotlinx-cbor-js:${Version.kotlin.ktor}"
+        const val websockets = "io.ktor:ktor-client-websockets-js:${Version.kotlin.ktor}"
+        const val contentNegotiation = "io.ktor:ktor-client-content-negotiation-js:${Version.kotlin.ktor}"
+    }
+
+    val native = Native
+
+    object Native {
+        const val client = "io.ktor:ktor-client-curl:${Version.kotlin.ktor}"
+        const val json = "io.ktor:ktor-serialization-kotlinx-json-native:${Version.kotlin.ktor}"
+        const val xml = "io.ktor:ktor-serialization-kotlinx-xml-native:${Version.kotlin.ktor}"
+        const val cbor = "io.ktor:ktor-serialization-kotlinx-cbor-native:${Version.kotlin.ktor}"
+        const val websockets = "io.ktor:ktor-client-websockets-native:${Version.kotlin.ktor}"
+        const val contentNegotiation = "io.ktor:ktor-client-content-negotiation-native:${Version.kotlin.ktor}"
+    }
+
+    // Logger
+    const val logger = "io.ktor:ktor-client-logging:${Version.kotlin.ktor}"
+
+    // Testing
+    const val mock = "io.ktor:ktor-client-mock:${Version.kotlin.ktor}"
+}
