+++ antibytes-dependency/src/main/kotlin/tech/antibytes/gradle/dependency/packages/kmp/Test.kt
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
+ *
+ * Use of this source code is governed by Apache License, Version 2.0
+ */
+
+package tech.antibytes.gradle.dependency.packages.kmp
+
+import tech.antibytes.gradle.dependency.Version
+
+object Test {
+    const val common = "org.jetbrains.kotlin:kotlin-test-common:${Version.kotlin.stdlib}"
+    const val coroutines = "org.jetbrains.kotlinx:kotlinx-coroutines-test:${Version.kotlin.coroutines}"
+    const val annotations = "org.jetbrains.kotlin:kotlin-test-annotations-common:${Version.kotlin.stdlib}"
+    const val js = "org.jetbrains.kotlin:kotlin-test-js:${Version.kotlin.stdlib}"
+    const val jvm = "org.jetbrains.kotlin:kotlin-test:${Version.kotlin.stdlib}"
+    const val junit = "org.jetbrains.kotlin:kotlin-test-junit:${Version.kotlin.stdlib}"
+}
