+++ antibytes-dependency/src/main/kotlin/tech/antibytes/gradle/dependency/version/Js.kt
@@ -1,0 +1,19 @@
+/*
+ * Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
+ *
+ * Use of this source code is governed by Apache License, Version 2.0
+ */
+
+package tech.antibytes.gradle.dependency.version
+
+object Js {
+    /**
+     * [KotlinNodeJS](https://github.com/Kotlin/kotlinx-nodejs)
+     */
+    const val nodeJs = "0.0.7"
+
+    /**
+     * [NodeJs](https://nodejs.org/en/)
+     */
+    const val node = "18.9.0"
+}
